Rename general.py to vnv_wrapper.py and fix the corresponding imports
